+++ apps/src/code-studio/pd/form_components_func/FormComponent.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export const FormContext = React.createContext({});
+
+/**
+ * @param {String} name
+ * @param {Array<String>} errors
+ * @returns {String|undefined}
+ */
+export const getValidationState = (name, errors) => {
+  if (errors.includes(name)) {
+    return 'error';
+  }
+};
