+++ apps/src/code-studio/pd/form_components_func/form/UsPhoneNumberInput.jsx
@@ -1,0 +1,26 @@
+import PropTypes from 'prop-types';
+import React, {useContext} from 'react';
+import UsPhoneNumberInputComponent from '../../form_components/UsPhoneNumberInput';
+import {getValidationState, FormContext} from '../FormComponent';
+/**
+ * Construct a controlled US phone number input
+ */
+export const UsPhoneNumberInput = props => {
+  const {errors, errorMessages, onChange, data} = useContext(FormContext);
+  const {name} = props;
+  return (
+    <UsPhoneNumberInputComponent
+      name={name}
+      validationState={getValidationState(name, errors)}
+      errorMessage={errorMessages[name]}
+      onChange={onChange}
+      value={data[name]}
+      {...props}
+    />
+  );
+};
+UsPhoneNumberInput.propTypes = {
+  ...UsPhoneNumberInputComponent.propTypes,
+  // the name of the input
+  name: PropTypes.string.isRequired
+};
