+++ apps/src/code-studio/pd/form_components_func/labeled/LabeledRadioButtons.jsx
@@ -1,0 +1,53 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+import {
+  ButtonsFromOptions,
+  ButtonsWithAdditionalTextFields,
+  ButtonsWithAdditionalTextFieldsFromOptions
+} from '../form/Buttons';
+import {defaultOptions} from '../LabeledFormComponent';
+
+export const LabeledRadioButtons = props => {
+  const passProps = {
+    ...defaultOptions(props.name, props.label),
+    type: 'radio',
+    ...props
+  };
+  return <ButtonsFromOptions {...passProps} />;
+};
+LabeledRadioButtons.propTypes = {
+  name: PropTypes.string.isRequired,
+  label: PropTypes.string
+};
+
+export const LabeledRadioButtonsWithAdditionalTextFields = props => {
+  const passProps = {
+    ...defaultOptions(props.name, props.label),
+    type: 'radio',
+    textFieldMap: props.textFieldMap,
+    ...props
+  };
+  return <ButtonsWithAdditionalTextFieldsFromOptions {...passProps} />;
+};
+LabeledRadioButtonsWithAdditionalTextFields.propTypes = {
+  name: PropTypes.string.isRequired,
+  label: PropTypes.string,
+  textFieldMap: PropTypes.object
+};
+
+export const LabeledDynamicRadioButtonsWithAdditionalTextFields = props => {
+  const passProps = {
+    ...defaultOptions(props.name, props.label),
+    type: 'radio',
+    options: props.options,
+    textFieldMap: props.textFieldMap,
+    ...props
+  };
+  return <ButtonsWithAdditionalTextFields {...passProps} />;
+};
+LabeledDynamicRadioButtonsWithAdditionalTextFields.propTypes = {
+  name: PropTypes.string.isRequired,
+  label: PropTypes.string,
+  options: PropTypes.array,
+  textFieldMap: PropTypes.object
+};
