+++ apps/src/code-studio/pd/form_components_func/labeled/LabeledSelect.jsx
@@ -1,0 +1,17 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+import {SelectFieldGroupFromOptions} from '../form/SelectFieldGroup';
+import {defaultOptions} from '../LabeledFormComponent';
+
+export const LabeledSelect = (props = {}) => {
+  const passProps = {
+    ...defaultOptions(props.name, props.label),
+    type: 'select',
+    ...props
+  };
+  return <SelectFieldGroupFromOptions {...passProps} />;
+};
+LabeledSelect.propTypes = {
+  name: PropTypes.string.isRequired,
+  label: PropTypes.string
+};
